Got streetTacos section working
streetTacos section is working responsively. decided to section that off and get things done from the top down. will now make a section for
Details, Ingredients, and Directions to work responsively as well. I like to know the section I'm working on is finished before starting to markup the next section. It might be able to be made a simpler way in the future, but for the purpose of this assignment, I want to make sure one thing works before moving onto the next. I could change to using grid area in the future to make it with less code, but for now this section works and I'm moving on.