Finished Footer, Socials, Sides
I got everything working (responsively) and within parameters. Added the hover to lists so you can keep track of your spot with a mouse. I am very close to the template, and had some good experienced and difficulties along the way. Alas, it is 6:30am and I might get time to take a final look at it tomorrow.