Fixed Print media Quiery Fonts
previously in body, that didn't correct for the specific places fonts were assigned before. Also changed line height for readability. known error: upon resizing the window to be smaller in safari browser, the search bar will try to not go back to it's original placing. if you refresh (or start from small view widths), it will be in the correct place. I tried a workaround. Isn't a problem on chrome